feature(core): add support for icon at submit button
