+++ modules/@ergonode/category-trees/src/components/Tabs/CategoryTreeDesignerTab.vue
@@ -23,53 +23,38 @@
                 :scope="scope"
                 :change-values="changeValues"
                 :errors="errors" />
-            <Button
-                :title="$t('core.buttons.submit')"
-                :floating="saveChangesButtonFloatingStyle"
-                @click.native="onSubmit">
-                <template
-                    v-if="isSubmitting"
-                    #prepend="{ color }">
-                    <IconSpinner :fill-color="color" />
-                </template>
-            </Button>
+            <UpdateCategoryTreeDesignerButton
+                :scope="scope"
+                :change-values="changeValues"
+                :errors="errors" />
         </template>
     </GridViewTemplate>
 </template>
 
 <script>
 import {
-    ALERT_TYPE,
-} from '@Core/defaults/alerts';
-import {
     DRAGGED_ELEMENT,
 } from '@Core/defaults/grid';
 import tabFeedbackMixin from '@Core/mixins/tab/tabFeedbackMixin';
+import UpdateCategoryTreeDesignerButton from '@Trees/components/Buttons/UpdateCategoryTreeDesignerButton';
 import CategoryTreeWrapper from '@Trees/components/CategoryTreeDesigner/CategoryTreeWrapper';
 import PRIVILEGES from '@Trees/config/privileges';
 import {
     GRAPHITE_LIGHT,
 } from '@UI/assets/scss/_js-variables/colors.scss';
-import {
-    Z_INDEX_LVL_2,
-} from '@UI/assets/scss/_js-variables/indexes.scss';
-import Button from '@UI/components/Button/Button';
 import DropZone from '@UI/components/DropZone/DropZone';
 import IconRemoveFilter from '@UI/components/Icons/Actions/IconRemoveFilter';
-import IconSpinner from '@UI/components/Icons/Feedback/IconSpinner';
 import GridViewTemplate from '@UI/components/Layout/Templates/GridViewTemplate';
 import VerticalTabBar from '@UI/components/TabBar/VerticalTabBar';
 import FadeTransition from '@UI/components/Transitions/FadeTransition';
 import {
-    mapActions,
     mapState,
 } from 'vuex';
 
 export default {
     name: 'CategoryTreeDesignerTab',
     components: {
-        Button,
-        IconSpinner,
+        UpdateCategoryTreeDesignerButton,
         VerticalTabBar,
         CategoryTreeWrapper,
         GridViewTemplate,
@@ -80,11 +65,6 @@ export default {
     mixins: [
         tabFeedbackMixin,
     ],
-    data() {
-        return {
-            isSubmitting: false,
-        };
-    },
     computed: {
         ...mapState('draggable', [
             'isElementDragging',
@@ -107,48 +87,8 @@ export default {
         isDropZoneVisible() {
             return this.isElementDragging === DRAGGED_ELEMENT.TEMPLATE;
         },
-        saveChangesButtonFloatingStyle() {
-            return {
-                bottom: '24px',
-                right: '24px',
-                zIndex: Z_INDEX_LVL_2,
-            };
-        },
         graphiteLightColor() {
             return GRAPHITE_LIGHT;
-        },
-    },
-    methods: {
-        ...mapActions('categoryTree', [
-            'updateCategoryTree',
-        ]),
-        onSubmit() {
-            if (this.isSubmitting) {
-                return;
-            }
-            this.isSubmitting = true;
-
-            this.removeScopeErrors(this.scope);
-            this.updateCategoryTree({
-                scope: this.scope,
-                onSuccess: this.onUpdateSuccess,
-                onError: this.onUpdateError,
-            });
-        },
-        onUpdateSuccess() {
-            this.$addAlert({
-                type: ALERT_TYPE.SUCCESS,
-                message: 'Category tree has been updated',
-            });
-
-            this.isSubmitting = false;
-
-            this.markChangeValuesAsSaved(this.scope);
-        },
-        onUpdateError(errors) {
-            this.onError(errors);
-
-            this.isSubmitting = false;
         },
     },
 };
