+++ modules/@ergonode/core/src/components/Tabs/LanguagesSettingsTab.vue
@@ -20,43 +20,27 @@
         </template>
         <template #grid>
             <LanguagesTreeWrapper />
-            <Button
-                :title="$t('core.buttons.submit')"
-                :floating="{ bottom: '24px', right: '24px' }"
-                @click.native="onSubmit">
-                <template
-                    v-if="isSubmitting"
-                    #prepend="{ color }">
-                    <IconSpinner :fill-color="color" />
-                </template>
-            </Button>
+            <UpdateLanguagesInheritanceButton
+                :scope="scope"
+                :change-values="changeValues"
+                :errors="errors" />
         </template>
     </GridViewTemplate>
 </template>
 
 <script>
+import UpdateLanguagesInheritanceButton from '@Core/components/Buttons/UpdateLanguagesInheritanceButton';
 import LanguagesTreeWrapper from '@Core/components/LanguagesTreeDesigner/LanguagesTreeWrapper';
 import PRIVILEGES from '@Core/config/privileges';
-import {
-    ALERT_TYPE,
-} from '@Core/defaults/alerts';
 import {
     DRAGGED_ELEMENT,
 } from '@Core/defaults/grid';
 import tabFeedbackMixin from '@Core/mixins/tab/tabFeedbackMixin';
 import {
-    getMappedTreeData,
-} from '@Core/models/mappers/languageTreeMapper';
-import {
-    isEmpty,
-} from '@Core/models/objectWrapper';
-import {
     GRAPHITE_LIGHT,
 } from '@UI/assets/scss/_js-variables/colors.scss';
-import Button from '@UI/components/Button/Button';
 import DropZone from '@UI/components/DropZone/DropZone';
 import IconRemoveFilter from '@UI/components/Icons/Actions/IconRemoveFilter';
-import IconSpinner from '@UI/components/Icons/Feedback/IconSpinner';
 import GridViewTemplate from '@UI/components/Layout/Templates/GridViewTemplate';
 import VerticalTabBar from '@UI/components/TabBar/VerticalTabBar';
 import FadeTransition from '@UI/components/Transitions/FadeTransition';
@@ -68,12 +52,11 @@ import {
 export default {
     name: 'LanguagesSettingsTab',
     components: {
+        UpdateLanguagesInheritanceButton,
         GridViewTemplate,
-        Button,
         IconRemoveFilter,
         DropZone,
         FadeTransition,
-        IconSpinner,
         LanguagesTreeWrapper,
         VerticalTabBar,
     },
@@ -107,15 +90,7 @@ export default {
         store.dispatch('gridDesigner/setGridData', treeToSet);
         store.dispatch('gridDesigner/setFullGridData', treeToSet);
     },
-    data() {
-        return {
-            isSubmitting: false,
-        };
-    },
     computed: {
-        ...mapState('gridDesigner', [
-            'fullGridData',
-        ]),
         ...mapState('draggable', [
             'isElementDragging',
         ]),
@@ -147,60 +122,6 @@ export default {
         ...mapActions('gridDesigner', {
             __clearGridDesignerStorage: '__clearStorage',
         }),
-        ...mapActions('core', [
-            'setLanguageTree',
-            'setDefaultLanguage',
-            'updateLanguageTree',
-        ]),
-        ...mapActions('authentication', [
-            'getUser',
-        ]),
-        onSubmit() {
-            if (this.isSubmitting) {
-                return;
-            }
-
-            if (!isEmpty(this.fullGridData)) {
-                this.isSubmitting = true;
-
-                const [
-                    languages,
-                ] = getMappedTreeData(this.fullGridData);
-
-                this.removeScopeErrors(this.scope);
-
-                this.updateLanguageTree({
-                    languages,
-                    onSuccess: () => this.onUpdateSuccess(languages),
-                    onError: this.onUpdateError,
-                });
-            }
-        },
-        async onUpdateSuccess(languages) {
-            this.setLanguageTree(languages);
-
-            await this.getUser({
-                onSuccess: this.onGetUserSuccess,
-            });
-        },
-        onUpdateError(message) {
-            this.$addAlert({
-                type: ALERT_TYPE.ERROR,
-                message,
-            });
-            this.isSubmitting = false;
-        },
-        onGetUserSuccess() {
-            this.setDefaultLanguage();
-
-            this.$addAlert({
-                type: ALERT_TYPE.SUCCESS,
-                message: 'Languages tree has been updated',
-            });
-            this.isSubmitting = false;
-
-            this.markChangeValuesAsSaved(this.scope);
-        },
     },
 };
 </script>
