+++ modules/@ergonode/segments/src/components/Tabs/ConditionDesignerTab.vue
@@ -24,16 +24,10 @@
                 :change-values="changeValues"
                 :errors="errors"
                 :disabled="!isAllowedToUpdate" />
-            <Button
-                :title="$t('core.buttons.submit')"
-                :floating="saveChangesButtonFloatingStyle"
-                @click.native="onSubmit">
-                <template
-                    v-if="isSubmitting"
-                    #prepend="{ color }">
-                    <IconSpinner :fill-color="color" />
-                </template>
-            </Button>
+            <UpdateSegmentConditionDesignerButton
+                :scope="scope"
+                :change-values="changeValues"
+                :errors="errors" />
         </template>
     </GridViewTemplate>
 </template>
@@ -41,23 +35,16 @@
 <script>
 import ConditionSetWrapper from '@Conditions/components/ConditionSetDesigner/ConditionSetWrapper';
 import {
-    ALERT_TYPE,
-} from '@Core/defaults/alerts';
-import {
     DRAGGED_ELEMENT,
 } from '@Core/defaults/grid';
 import tabFeedbackMixin from '@Core/mixins/tab/tabFeedbackMixin';
+import UpdateSegmentConditionDesignerButton from '@Segments/components/Buttons/UpdateSegmentConditionDesignerButton';
 import PRIVILEGES from '@Segments/config/privileges';
 import {
     GRAPHITE_LIGHT,
 } from '@UI/assets/scss/_js-variables/colors.scss';
-import {
-    Z_INDEX_LVL_2,
-} from '@UI/assets/scss/_js-variables/indexes.scss';
-import Button from '@UI/components/Button/Button';
 import DropZone from '@UI/components/DropZone/DropZone';
 import IconRemoveFilter from '@UI/components/Icons/Actions/IconRemoveFilter';
-import IconSpinner from '@UI/components/Icons/Feedback/IconSpinner';
 import GridViewTemplate from '@UI/components/Layout/Templates/GridViewTemplate';
 import VerticalTabBar from '@UI/components/TabBar/VerticalTabBar';
 import FadeTransition from '@UI/components/Transitions/FadeTransition';
@@ -69,8 +56,7 @@ import {
 export default {
     name: 'ConditionDesignerTab',
     components: {
-        IconSpinner,
-        Button,
+        UpdateSegmentConditionDesignerButton,
         GridViewTemplate,
         IconRemoveFilter,
         DropZone,
@@ -94,9 +80,6 @@ export default {
         };
     },
     computed: {
-        ...mapState('segment', [
-            'conditionSetId',
-        ]),
         ...mapState('draggable', [
             'isElementDragging',
         ]),
@@ -117,13 +100,6 @@ export default {
                 },
             ];
         },
-        saveChangesButtonFloatingStyle() {
-            return {
-                bottom: '24px',
-                right: '24px',
-                zIndex: Z_INDEX_LVL_2,
-            };
-        },
         isDropZoneVisible() {
             return this.isElementDragging === DRAGGED_ELEMENT.TEMPLATE;
         },
@@ -136,66 +112,12 @@ export default {
         this.__clearConditionStorage();
     },
     methods: {
-        ...mapActions('condition', [
-            'createConditionSet',
-            'updateConditionSet',
-        ]),
         ...mapActions('gridDesigner', {
             __clearGridDesignerStorage: '__clearStorage',
         }),
         ...mapActions('condition', {
             __clearConditionStorage: '__clearStorage',
         }),
-        ...mapActions('segment', [
-            'updateSegment',
-            '__setState',
-        ]),
-        onSubmit() {
-            if (this.isSubmitting) {
-                return;
-            }
-            this.isSubmitting = true;
-
-            this.removeScopeErrors(this.scope);
-
-            if (!this.conditionSetId) {
-                this.createConditionSet({
-                    scope: this.scope,
-                    onSuccess: this.onUpdateSuccess,
-                    onError: this.onUpdateError,
-                });
-            } else {
-                this.updateConditionSet({
-                    scope: this.scope,
-                    onSuccess: this.onUpdateSuccess,
-                    onError: this.onUpdateError,
-                });
-            }
-        },
-        async onUpdateSuccess(id) {
-            this.__setState({
-                key: 'conditionSetId',
-                value: id,
-            });
-
-            await this.updateSegment({
-                scope: this.scope,
-            });
-
-            this.$addAlert({
-                type: ALERT_TYPE.SUCCESS,
-                message: 'Segment conditions have been updated',
-            });
-
-            this.isSubmitting = false;
-
-            this.markChangeValuesAsSaved(this.scope);
-        },
-        onUpdateError(errors) {
-            this.onError(errors);
-
-            this.isSubmitting = false;
-        },
     },
 };
 </script>
